Move ROS file server loader into the ROS backend package
