+++ docs/examples/02_description_models/files/02_robot_from_ros_with_cache.py
@@ -3,8 +3,7 @@ import os
 import compas
 from compas.robots import RobotModel
 from compas_fab.backends import RosClient
-from compas_fab.robots import RobotSemantics
-from compas_fab.robots import RosFileServerLoader
+from compas_fab.backends import RosFileServerLoader
 
 # Set high precision to import meshes defined in meters
 compas.PRECISION = '12f'
