+++ src/compas_fab/robots/__init__.py
@@ -50,6 +50,5 @@ from .planning_scene import *
 from .units import *
 from .robot import *
 from .semantics import *
-from .ros_fileserver_loader import *
 
 __all__ = [name for name in dir() if not name.startswith('_')]
